Move EntryEditor classes to GUI package
