+++ src/main/java/net/sf/jabref/gui/GenFieldsCustomizer.java
@@ -23,6 +23,7 @@ import javax.swing.*;
 import javax.swing.border.TitledBorder;
 
 import net.sf.jabref.*;
+import net.sf.jabref.gui.entryeditor.EntryEditorTabList;
 import net.sf.jabref.gui.help.HelpAction;
 
 import com.jgoodies.forms.builder.ButtonBarBuilder;
