+++ src/main/java/net/sf/jabref/gui/entryeditor/EntryEditor.java
@@ -1,4 +1,4 @@
-/*  Copyright (C) 2003-2014 JabRef contributors.
+/*  Copyright (C) 2003-2015 JabRef contributors.
     This program is free software; you can redistribute it and/or modify
     it under the terms of the GNU General Public License as published by
     the Free Software Foundation; either version 2 of the License, or
@@ -13,7 +13,7 @@
     with this program; if not, write to the Free Software Foundation, Inc.,
     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
-package net.sf.jabref;
+package net.sf.jabref.gui.entryeditor;
 
 import java.awt.AWTKeyStroke;
 import java.awt.BorderLayout;
@@ -49,6 +49,7 @@ import javax.swing.event.ChangeEvent;
 import javax.swing.event.ChangeListener;
 import javax.swing.text.JTextComponent;
 
+import net.sf.jabref.*;
 import net.sf.jabref.gui.fieldeditors.*;
 import net.sf.jabref.logic.autocompleter.AutoCompleter;
 import net.sf.jabref.export.LatexFieldFormatter;
