+++ src/main/java/net/sf/jabref/gui/entryeditor/EntryEditorTabList.java
@@ -13,7 +13,10 @@
     with this program; if not, write to the Free Software Foundation, Inc.,
     51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 */
-package net.sf.jabref;
+package net.sf.jabref.gui.entryeditor;
+
+import net.sf.jabref.Globals;
+import net.sf.jabref.JabRefPreferences;
 
 import java.util.ArrayList;
 import java.util.Arrays;
