+++ src/main/java/net/sf/jabref/gui/fieldeditors/FileListEditor.java
@@ -46,6 +46,7 @@ import net.sf.jabref.gui.AutoCompleteListener;
 import net.sf.jabref.gui.FileListEntry;
 import net.sf.jabref.gui.FileListEntryEditor;
 import net.sf.jabref.gui.FileListTableModel;
+import net.sf.jabref.gui.entryeditor.EntryEditor;
 import net.sf.jabref.logic.l10n.Localization;
 import net.sf.jabref.logic.util.StringUtil;
 import net.sf.jabref.util.JabRefDesktop;
