+++ src/test/java/net/sf/jabref/logic/autocompleter/AutoCompleterTest.java
@@ -1,6 +1,7 @@
 package net.sf.jabref.logic.autocompleter;
 
 import net.sf.jabref.*;
+import net.sf.jabref.gui.entryeditor.EntryEditor;
 import net.sf.jabref.gui.fieldeditors.FieldEditor;
 import net.sf.jabref.gui.fieldeditors.FieldTextArea;
 import net.sf.jabref.imports.ParserResult;
